finish copycat unittest
