+++ tests/src/base_2.py
@@ -1,0 +1,1 @@
+print('Hello,', input())
