+++ tests/test_submission.py
@@ -25,138 +25,6 @@ S_NAMES = {
 }
 
 
-@pytest.fixture
-def save_source(tmp_path):
-    def save_source(filename, source, lang, ext=None):
-        '''
-        save user source codes to tmp dir
-        currently only support one file.
-
-        Args:
-            filename: the source code's filename without extension
-            source: source code
-            lang: programming language, only accept {0, 1, 2}
-            ext: main script extension want to use, if None, decided by lang
-
-        Returns:
-            a zip file contains source code
-        '''
-        if not ext:
-            ext = ['.c', '.cpp', '.py'][lang]
-
-        # set path
-        name = tmp_path / (filename + ext)
-        zip_path = tmp_path / f'{name}.zip'
-
-        if name.exists():
-            raise FileExistsError(name)
-
-        with open(name, 'w') as f:
-            f.write(source)
-        with ZipFile(zip_path, 'w') as f:
-            f.write(name, arcname=f'main{ext}')
-        return True
-
-    return save_source
-
-
-@pytest.fixture
-def get_source(tmp_path):
-    def get_source(filename):
-        '''
-        get users zipped source by filename
-
-        Args:
-            filename: a string denote the source code's filename include extension
-        
-        Returns:
-            the zip file
-        '''
-        path = tmp_path / f'{filename}.zip'
-
-        if not path.exists():
-            raise FileNotFoundError(path)
-
-        return open(path, 'rb')
-
-    return get_source
-
-
-@pytest.fixture
-def submit_once(forge_client, get_source):
-    def submit_once(name, pid, filename, lang, client=None):
-        _client = forge_client(name) if client is None else client
-        now = datetime.utcnow()
-        rv, rv_json, rv_data = BaseTester.request(
-            _client,
-            'post',
-            '/submission',
-            json={
-                'problemId': pid,
-                'languageType': lang
-            },
-        )
-        assert rv.status_code == 200, rv_json
-
-        submission_id = rv_data['submissionId']
-        token = rv_data['token']
-
-        submission = {
-            'submissionId': submission_id,
-            'problemId': pid,
-            'username': name,
-            'lang': lang,
-            'timestamp': now
-        }
-
-        files = {
-            'code': (
-                get_source(filename),
-                f'main.{["c", "cpp", "py"][lang]}',
-            )
-        }
-        rv, rv_json, rv_data = BaseTester.request(
-            _client,
-            'put',
-            f'/submission/{submission_id}?token={token}',
-            data=files,
-        )
-        assert rv.status_code == 200, rv_json
-
-        if client is None:
-            _client.cookie_jar.clear()
-        return submission_id
-
-    return submit_once
-
-
-@pytest.fixture
-def submit(submit_once, forge_client):
-    def submit(names, pids, count, filename='base.c', lang=0):
-        n2p = defaultdict(list)  # name to pid
-
-        for n, p, _ in zip(names, pids, 'x' * count):
-            n2p[n].append(p)
-
-        n2s = defaultdict(list)  # name to submission id
-        for name, ps in n2p.items():
-            client = forge_client(name)
-            for p in ps:
-                n2s[name].append(
-                    submit_once(
-                        name=name,
-                        pid=p,
-                        filename=filename,
-                        lang=lang,
-                        client=client,
-                    ))
-            client.cookie_jar.clear()
-
-        return n2s
-
-    return submit
-
-
 @pytest.fixture(autouse=True)
 def submission_testcase_setup(
     save_source,
